# backend/docs/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "service.ChangePasswordRequest": {
        "properties": {
          "newPassword": {
            "minLength": 8,
            "type": "string"
          },
          "oldPassword": {
            "type": "string"
          }
        },
        "required": [
          "newPassword",
          "oldPassword"
        ],
        "type": "object"
      },
...
      "service.RegisterRequest": {
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "minLength": 8,
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "username"
        ],
        "type": "object"
      },
...
      "service.UpdateUserRequest": {
        "properties": {
          "departmentId": {
            "type": "integer"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "输入格式: Bearer {session_id}，session_id 从登录接口获取 (Swagger 暂不支持 Cookie 自动发送，请手动输入)",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "email": "support@example.com",
      "name": "API Support",
      "url": "http://www.example.com/support"
    },
    "description": "DevOps 运维平台接口文档，登录后在 Authorize 中输入 Bearer {session_id} 即可调用接口",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService": "http://swagger.io/terms/",
    "title": "DevOps 运维平台 API",
    "version": "1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/user/assign-roles": {
      "post": {
        "description": "为用户分配角色",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "description": "分配信息 {userId: 1, roleIds: [1,2,3]}",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "分配成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "参数错误"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "分配角色",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/user/change-password": {
      "post": {
        "description": "修改当前用户密码",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/service.ChangePasswordRequest"
              }
            }
          },
          "description": "密码修改信息",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "修改成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "参数错误"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "未认证"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "修改密码",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/user/delete": {
      "post": {
        "description": "删除指定用户",
        "parameters": [
          {
            "description": "用户ID",
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "删除成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "参数错误"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "删除用户",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/user/detail": {
      "get": {
        "description": "根据ID获取用户详细信息",
        "parameters": [
          {
            "description": "用户ID",
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "用户详情"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "参数错误"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "用户不存在"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "获取用户详情",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/user/info": {
      "get": {
        "description": "获取当前登录用户的详细信息",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "用户信息"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "未认证"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "获取当前用户信息",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/user/list": {
      "get": {
        "description": "获取用户列表（支持分页和关键词搜索）",
        "parameters": [
          {
            "description": "页码",
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "type": "integer"
            }
          },
          {
            "description": "每页数量",
            "in": "query",
            "name": "pageSize",
            "schema": {
              "default": 10,
              "type": "integer"
            }
          },
          {
            "description": "搜索关键词（用户名、姓名、邮箱）",
            "in": "query",
            "name": "keyword",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "用户列表"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "获取用户列表",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/user/lock": {
      "post": {
        "description": "锁定指定用户",
        "parameters": [
          {
            "description": "用户ID",
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "锁定成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "参数错误"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "锁定用户",
        "tags": [
          "用户管理"
        ]
      }
    },
...
    "/user/permissions": {
      "get": {
        "description": "获取当前登录用户的权限编码集合",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "用户权限"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "未认证"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "获取当前用户权限",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/user/register": {
      "post": {
        "description": "注册新用户（仅限本地认证）",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/service.RegisterRequest"
              }
            }
          },
          "description": "注册信息",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "注册成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "参数错误"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "服务器错误"
          }
        },
        "summary": "用户注册",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/user/reset-password": {
      "post": {
        "description": "管理员重置用户密码",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "description": "重置信息 {userId: 1, newPassword: 'xxx'}",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "重置成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "参数错误"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "重置密码（管理员）",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/user/unlock": {
      "post": {
        "description": "解锁指定用户",
        "parameters": [
          {
            "description": "用户ID",
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "解锁成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "参数错误"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "解锁用户",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/user/update": {
      "post": {
        "description": "更新用户的基本信息",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/service.UpdateUserRequest"
              }
            }
          },
          "description": "用户信息",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "更新成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "参数错误"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "更新用户信息",
        "tags": [
          "用户管理"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1"
    },
    {
      "url": "https://localhost:8000/api/v1"
    }
  ]
}